Split out common methods for other projects
`pobinterface.lua` is now intended to be shared across projects so comtains generally useful functions.

`commands.lua` contains only command methods that will be called from Python.

Also fixes an issue with the updater which failed to reselect the correct skill if gems moved and PoB selected a group with a non-active skill.